icu_congestion/requirements.txt
# icu_congestion/requirements.txt
numpy==2.2.5
pandas==2.2.3
scikit-learn==1.6.1
imbalanced-learn==0.13.0
catboost==1.2.8
SQLAlchemy==2.0.40
APScheduler==3.11.0
uvicorn==0.34.2
fastapi>=0.115.12
python-dotenv==1.1.0
pymysql==1.1.1
